+++ src/main/java/ru/maxima/bankalgorithm2/models/Person.java
@@ -3,13 +3,15 @@ package ru.maxima.bankalgorithm2.models;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class Person {
 
     private String name;
-    private Double wallet;
-    private Double appendFromBank;
+    private BigDecimal wallet;
+    private BigDecimal appendFromBank;
 
     @Override
     public String toString() {
